Provide message that build has been cancelled but the log may still be processing
Based on how things are currently setup, there would need to be a lot
of rework to keep polling and handle scroll events for builds that were
cancelled but are still processing the log. The window is small where
the polling would be beneficial, so instead of a big rewrite we will
display a message that the log is still processing and they can refresh
to get the latest log results.