Fix interprojectbuild finishing race condition
Handling the marking of an ipb as finished in the repository build visitor had an unforeseen consequence.
because it uses a different queue than the modulebuild visitors, so builds are being marked as successful when really there is a failed module that has yet to be marked as failed. Moved the logic for marking a ipb as success into the ipb module build visitor to resolve this.