Fail empty branch builds
Builds without modules are "technically" successful in taking no
action. This is usually due to no modules being picked up by discovery
due to malformed files etc. It is confusing for maintainers and developers
to see something that did not take the _expected_ action to terminate
successfully. This change fails branch builds that have not built anything
to be considered a failure.