Change Whitelist behavior
For the testing of this in QA we need to have only certain projects
build, but since things are mostly opted-in now we don't want to || with
whether a project has a `.blazar-enabled` or `.blazar.yaml`.

This makes Blazar ONLY handle git-events for things in the white list,
if the white-list is non-emtpy. Otherwise it will check for the
existance of the files we usually use to determine if a project is `On
Blazar`.