Fix Inter Project Build commit Resolution
If an InterProject build comes just after a push to a repository, but
before the push build is queued it can prevent Blazar from building
the changed modules correctly.

This change makes Blazar use the last built commit for InterProject
builds so that they behave like a "rebuild to pick up dependencies",
instead of having them pick up new commits and not build the right
modules.